+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/FilmDbStorage.java
@@ -213,7 +213,7 @@ public class FilmDbStorage implements FilmStorage {
 	}
 
 	@Override
-	public void deleteFilm(Integer filmId) {
+	public void deleteFilm(Long filmId) {
 		final String sqlQuery = "DELETE FROM films WHERE FILM_ID = ?";
 		jdbcTemplate.update(sqlQuery, filmId);
 	}
